Extract state management into a hook
