+++ src/components/filterform/FilterField.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react'
 import { Select } from '../select'
-import { useFormikContext } from 'formik'
-import { FilterForm } from './FiltersForm'
+import { useFilterField } from './useFilterField'
 
 type FilterType = 'select'
 
@@ -28,66 +27,6 @@ type EmptyFilter = {
 
 export type Filter = SelectFilter | EmptyFilter
 
-const TypeToOperatorOptions = {
-  select: [
-    { label: 'is any of', value: 'equals' },
-    { label: 'is not any of', value: 'notEqual' },
-    { label: 'has no value', value: 'blank' },
-    { label: 'has any value', value: 'notBlank' },
-  ],
-}
-
-type FilterFieldState = {
-  config: FilterConfig | undefined
-  filterConfigs: FilterConfig[]
-  operatorOptions: { label: string; value: string }[]
-  valueOptions: { label: string; value: string }[] | undefined
-}
-
-type Action = {
-  type: 'setField' | 'setOperator'
-  payload: string | undefined
-}
-
-const buildValueOptions = (config: FilterConfig, operator: string) => {
-  if (
-    config.type === 'select' &&
-    (operator === 'blank' || operator === 'notBlank')
-  ) {
-    return
-  }
-
-  return config.options
-}
-
-const filterFieldReducer = (state: FilterFieldState, action: Action) => {
-  switch (action.type) {
-    case 'setField': {
-      const config =
-        state.filterConfigs.find((config) => config.field === action.payload) ||
-        state.filterConfigs[0]
-
-      return {
-        ...state,
-        config,
-        operatorOptions: config ? TypeToOperatorOptions[config.type] : [],
-        valueOptions: config?.options,
-      }
-    }
-    case 'setOperator': {
-      if (!state.config) return state
-      if (!action.payload) return state
-
-      return {
-        ...state,
-        valueOptions: buildValueOptions(state.config, action.payload),
-      }
-    }
-    default:
-      return state
-  }
-}
-
 type FilterFieldProps = {
   filterConfigs: FilterConfig[]
   index: number
@@ -99,41 +38,13 @@ export const FilterField = ({
   index,
   filter,
 }: FilterFieldProps) => {
-  const { values, setFieldValue } = useFormikContext<FilterForm>()
-  const [{ config, operatorOptions, valueOptions }, dispatch] =
-    React.useReducer(
-      filterFieldReducer,
-      {
-        filterConfigs,
-        config: undefined,
-        operatorOptions: [],
-        valueOptions: [],
-      },
-      (state) => {
-        return filterFieldReducer(state, {
-          type: 'setField',
-          payload: filter.field,
-        })
-      }
-    )
-
-  const formKey = `filters.${index}`
-
-  const handleFieldChange = (value: string) => {
-    dispatch({
-      type: 'setField',
-      payload: value,
-    })
-    setFieldValue(`${formKey}.values`, [])
-  }
-
-  const handleOperatorChange = (value: string) => {
-    dispatch({
-      type: 'setOperator',
-      payload: value,
-    })
-    setFieldValue(`${formKey}.values`, undefined)
-  }
+  const {
+    formKey,
+    operatorOptions,
+    valueOptions,
+    handleFieldChange,
+    handleOperatorChange,
+  } = useFilterField({ filterConfigs, filter, index })
 
   const fieldOptions = React.useMemo(() => {
     return filterConfigs.map((filter) => ({
@@ -141,19 +52,6 @@ export const FilterField = ({
       value: filter.field,
     }))
   }, [filterConfigs])
-
-  React.useEffect(() => {
-    const currentValue = values.filters[index]
-
-    if (!currentValue.field) {
-      setFieldValue(`${formKey}.field`, config?.field)
-      setFieldValue(`${formKey}.type`, config?.type)
-    }
-
-    if (!currentValue.operator && operatorOptions.length > 0) {
-      setFieldValue(`${formKey}.operator`, operatorOptions[0].value)
-    }
-  }, [config])
 
   return (
     <div className="grid grid-cols-3 gap-2">
